pyproject.toml
[tool.poetry]
name = "ai-werewolf"
version = "0.1.0"
description = ""
authors = ["antoninduval <antonin@im-vitro.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.3.8"
chatgpt = "^2.2212.0"
faker = "^33.1.0"
termcolor = "^2.5.0"
openai = "^1.58.1"
transformers = "^4.47.1"
langchain-community = "^0.3.13"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.14"
torch = "^2.5.1"
loguru = "^0.7.3"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
